Extract SemesterBadge component
