+++ www/src/js/views/layout/GlobalSearch.scss
@@ -142,11 +142,6 @@ $icon-dist-lg: $input-width-open-lg - $input-width;
     flex: 0 0 auto;
     margin-right: 0.4rem;
   }
-
-  :global(.badge) {
-    margin-right: 0.2rem;
-    border-radius: 1px;
-  }
 }
 
 .noResults {
@@ -173,21 +168,4 @@ $icon-dist-lg: $input-width-open-lg - $input-width;
   composes: h4 from global;
   display: block;
   white-space: normal;
-}
-
-.sem1,
-.sem2,
-.sem3,
-.sem4 {
-  color: var(--body-color);
-}
-
-@each $semester in (1, 2, 3, 4) {
-  .sem#{$semester} {
-    background: darken($gray-lightest, 5% * $semester);
-
-    @include night-mode {
-      background: lighten(#2c2c2c, 5% * $semester);
-    }
-  }
 }
